Improve repository health and generated diffs
