+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# Changelog
+
+All notable user-facing changes should be recorded here.
+
+## Unreleased
+
+- Added agent-oriented CLI discovery, schemas, JSON/TOON envelopes, resumable jobs, and artifact surfaces.
+- Added first-class platform workflow commands for common project, assertion, incident, account, release, deployment, access, integration, protocol-manager, transfer, event, and search operations.
+- Added formatted generated dApp API client output for reviewable OpenAPI regeneration diffs.
+- Added repository-local CI, advisory audit, contributor, and security documentation.
