+++ Makefile
@@ -1,3 +1,25 @@
+.PHONY: ci fmt-check clippy test doc diff-check audit regenerate regenerate-dev
+
+ci: fmt-check clippy test doc diff-check
+
+fmt-check:
+	cargo fmt --all -- --check
+
+clippy:
+	cargo +nightly-2026-01-07 clippy --all-targets --workspace --locked --profile dev -- -D warnings -D clippy::pedantic
+
+test:
+	PCL_AUTH_NO_BROWSER=1 cargo test -q --workspace --all-targets
+
+doc:
+	cargo doc -q --workspace --no-deps
+
+diff-check:
+	git diff --check
+
+audit:
+	cargo deny check advisories
+
 # Regenerate dapp-api-client from latest OpenAPI spec
 regenerate:
 	@echo "Regenerating dapp-api-client from latest OpenAPI spec..."
