+++ crates/dapp-api-client/Cargo.toml
@@ -40,7 +40,9 @@ serde_json = { workspace = true }
 chrono = { workspace = true }
 anyhow = { workspace = true }
 openapiv3 = "2.2"
+prettyplease = "0.2"
 progenitor = "0.11.2"
+syn = { version = "2", features = ["full"] }
 
 [dev-dependencies]
 tokio = { workspace = true }
