+++ crates/pcl/phoundry/src/lib.rs
@@ -5,3 +5,5 @@ pub mod build_and_flatten;
 pub mod compile;
 pub mod error;
 pub mod phorge_test;
+
+pub const DEFAULT_ASSERTION_CONTRACTS_DIR: &str = "assertions/src";
