+++ deny.toml
@@ -1,0 +1,11 @@
+[advisories]
+# Keep the advisory gate explicit even while upstream Foundry/Alloy dependencies
+# carry known transitive advisories. Revisit these ignores during dependency
+# upgrade work; do not add new ignores without recording the upstream owner.
+ignore = [
+    { id = "RUSTSEC-2024-0436", reason = "Transitive through Foundry/Alloy; paste is unmaintained but no direct PCL usage." },
+    { id = "RUSTSEC-2024-0437", reason = "Transitive through Foundry/trezor protobuf dependency; tracked upstream." },
+    { id = "RUSTSEC-2026-0002", reason = "Transitive lru versions through Foundry/Alloy; tracked upstream." },
+    { id = "RUSTSEC-2026-0097", reason = "Transitive rand 0.8 through Foundry/Alloy; tracked upstream." },
+    { id = "RUSTSEC-2026-0104", reason = "Transitive rustls-webpki through reqwest/Alloy; tracked upstream." },
+]
